refactor: remove old services & utils
refactor: move popup options to components
refactor: unify signal with storage utility
