+++ components/options/Text.tsx
